+++ lib/src/components/painter/amount_chart/amount_bar_painter.dart
@@ -86,8 +86,8 @@ class AmountBarPainter extends BarPainter<_AmountBarItem> {
 
     final double intervalOfBars = size.width / dayCount;
     final int length = dataList.length;
-    final int viewLimitDay = getViewModeLimitDay(viewMode);
-    final dayFromScrollOffset = getDayFromScrollOffset();
+    final int viewLimitDay = viewMode.dayCount;
+    final dayFromScrollOffset = currentDayFromScrollOffset;
     final DateTime startDateTime = getBarRenderStartDateTime(dataList);
     final int startIndex = dataList.getLowerBound(startDateTime);