+++ lib/src/components/painter/time_chart/time_bar_painter.dart
@@ -173,9 +173,9 @@ class TimeBarPainter extends BarPainter<_TimeBarItem> {
     final int pivotHeight = pivotBottom > topHour ? pivotBottom - topHour : 24;
     final int length = dataList.length;
     final double height = size.height;
-    final int viewLimitDay = getViewModeLimitDay(viewMode);
+    final int viewLimitDay = viewMode.dayCount;
 
-    final int dayFromScrollOffset = getDayFromScrollOffset();
+    final int dayFromScrollOffset = currentDayFromScrollOffset;
     final DateTime startDateTime = getBarRenderStartDateTime(dataList);
     final int startIndex = dataList.getLowerBound(startDateTime);